+++ TripKitAndroid/src/androidTest/java/com/skedgo/android/tripkit/ServiceResponseTest.java
@@ -6,7 +6,7 @@ import com.google.gson.Gson;
 import com.google.gson.stream.JsonReader;
 import com.skedgo.android.common.model.Location;
 import skedgo.tripkit.routing.RealTimeVehicle;
-import com.skedgo.android.common.model.Shape;
+import skedgo.tripkit.routing.Shape;
 
 import java.io.IOException;
 import java.io.InputStream;
