+++ TripKitAndroid/src/main/java/com/skedgo/android/tripkit/ServiceResponse.java
@@ -4,7 +4,7 @@ import android.support.annotation.Nullable;
 
 import com.google.gson.annotations.JsonAdapter;
 import skedgo.tripkit.routing.RealTimeVehicle;
-import com.skedgo.android.common.model.Shape;
+import skedgo.tripkit.routing.Shape;
 
 import org.immutables.gson.Gson;
 import org.immutables.value.Value;
