+++ TripKitAndroid/src/main/java/com/skedgo/android/tripkit/TransitService.java
@@ -1,7 +1,7 @@
 package com.skedgo.android.tripkit;
 
 import skedgo.tripkit.routing.RealTimeVehicle;
-import com.skedgo.android.common.model.Shape;
+import skedgo.tripkit.routing.Shape;
 
 import org.immutables.gson.Gson;
 import org.immutables.value.Value;
